Update on "[WIP] [JIT] Support calling Tensor.element_size() in TorchScript"
# Problem

Resolves https://github.com/pytorch/pytorch/issues/33620

Differential Revision: [D7050644](https://our.internmc.facebook.com/intern/diff/D7050644/)

[ghstack-poisoned]
